Rename proposals to strategies
