+++ frontend/src/lib/components/StrategyForm.svelte
@@ -1,7 +1,8 @@
 <script lang="ts">
   import { onMount } from "svelte";
+  import { goto } from "$app/navigation";
   import Button from "$lib/components/Button.svelte";
-  import type { bestOptions } from "$lib/data/assetProposal";
+  import type { bestOptions } from "$lib/data/assetStrategy";
   import StratPieChart from "./StratPieChart.svelte";
   import { account } from "$lib/svark";
   import strats from "$lib/stores/strats";
@@ -52,6 +53,8 @@
         data: assets,
       },
     ]);
+
+    goto("/app/strategies");
   }
 </script>
 
@@ -108,7 +111,10 @@
 
   <div class="sticky top-2 h-fit">
     <StratPieChart {assets} />
-    <div class="mb-8">
+    <div
+      class="w-fit mx-auto flex flex-col gap-5
+     mt-6"
+    >
       {#if valid}
         <p>Looks good</p>
       {:else}
@@ -116,11 +122,11 @@
           >The total should be 100%. Right now it's {total}</p
         >
       {/if}
+      <Button
+        type="submit"
+        color="lightGreen">Submit</Button
+      >
     </div>
-    <Button
-      type="submit"
-      color="lightGreen">Submit</Button
-    >
   </div>
 </form>
