+++ frontend/src/lib/data/assetStrategy.js
